requirements.txt
aiosqlite
clerk-backend-api
colorlog
email-validator
fastapi
minio
openai
openai-agents
openai-chatkit
pillow
pydantic-settings
python-multipart
sqlalchemy
typer
uvicorn[standard]
